Test generation of docs
As part of the test suite it’s now tested whether the docs can be successfully generated. This should prevent the docs from breaking without anybody noticing. The test will fail on the first warning. To get a list of all warnings make the docs as usual with `make html`. This will succeed even in the presence of multiple warnings.

All dependencies for docs and testing are now part of the extra `dev`. The old extras have been removed to simplify the management of dependencies/extras.