Use intervaltree to back the label-list
Instead of storing the labels in a simple list, an interval-tree is now
used. The library used provides most of the functionality needed to do
operations like merging, finding overlaps and splitting. This reduces
the code complexity. In contrast to the previous approach the labels
have no fixed order anymore. Therefore it is important to sort them if
needed.